requirements.txt
# ============================================
# PYTHON VERSION REQUIRED: 3.9 - 3.11
# ============================================
# Recommended: Python 3.10 or 3.11
# Minimum: Python 3.9
# Not compatible with Python 3.12+ (TensorFlow issues)
# Not compatible with Python 3.8 or below

# ============================================
# CORE FLASK DEPENDENCIES (Required)
# ============================================
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# ============================================
# DATA PROCESSING (Required)
# ============================================
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# ============================================
# MACHINE LEARNING (Required)
# ============================================
tensorflow==2.15.0

# FOR APPLE SILICON (M1/M2/M3) MACS ONLY:
# Comment out tensorflow above and uncomment these:
# tensorflow-macos==2.15.0
# tensorflow-metal==1.1.0

# ============================================
# STOCK DATA FETCHING (Required)
# ============================================
yfinance==0.2.32
requests==2.31.0

# ============================================
# GOOGLE GEMINI AI (Required)
# ============================================
google-generativeai==0.3.1

# ============================================
# OPTIONAL DEPENDENCIES
# ============================================

# For Tor proxy support (if Yahoo Finance is blocked in your region)
requests[socks]==2.31.0

# For loading .env file (alternative to config.py)
python-dotenv==1.0.0

# ============================================
# DEVELOPMENT TOOLS (Optional - for developers only)
# ============================================
pytest==7.4.3       # For running tests
black==23.12.0      # For code formatting
flake8==6.1.0       # For code linting

# ============================================
# BUILT-IN PYTHON MODULES (No install needed)
# ============================================
# The following are built into Python:
# - sqlite3 (Database)
# - smtplib (Email sending)
# - email (Email handling)
# - hashlib (Password hashing)
# - json (JSON parsing)
# - datetime (Date/time handling)
# - os (Operating system interface)
# - typing (Type hints)

